Move all UI code to soccer/ui
