+++ soccer/CMakeLists.txt
@@ -9,9 +9,7 @@ include_directories(SYSTEM ${EIGEN_INCLUDE_DIR})
 # Note that main.cpp and LogViewer.cpp aren't present in this list because they're executable-specific
 set(ROBOCUP_LIB_SRC
     "BatteryProfile.cpp"
-    "BatteryWidget.cpp"
     "Configuration.cpp"
-    "FieldView.cpp"
     "gameplay/GameplayModule.cpp"
     "gameplay/robocup-py.cpp"
     "joystick/Joystick.cpp"
@@ -20,7 +18,6 @@ set(ROBOCUP_LIB_SRC
     "joystick/SpaceNavJoystick.cpp"
     "KickEvaluator.cpp"
     "Logger.cpp"
-    "MainWindow.cpp"
     "motion/MotionControl.cpp"
     "motion/TrapezoidalMotion.cpp"
     "NewRefereeModule.cpp"
@@ -46,17 +43,20 @@ set(ROBOCUP_LIB_SRC
     "planning/TargetVelPathPlanner.cpp"
     "planning/TrapezoidalPath.cpp"
     "Processor.cpp"
-    "ProtobufTree.cpp"
     "radio/SimRadio.cpp"
     "radio/USBRadio.cpp"
-    "RefereeTab.cpp"
     "Robot.cpp"
     "RobotConfig.cpp"
-    "RobotStatusWidget.cpp"
-    "RobotWidget.cpp"
-    "SimFieldView.cpp"
-    "StripChart.cpp"
     "SystemState.cpp"
+    "ui/BatteryWidget.cpp"
+    "ui/FieldView.cpp"
+    "ui/MainWindow.cpp"
+    "ui/ProtobufTree.cpp"
+    "ui/RefereeTab.cpp"
+    "ui/RobotStatusWidget.cpp"
+    "ui/RobotWidget.cpp"
+    "ui/SimFieldView.cpp"
+    "ui/StripChart.cpp"
     "ui/StyleSheetManager.cpp"
     "VisionReceiver.cpp"
     "vision/ball/BallBounce.cpp"
@@ -93,10 +93,10 @@ endif()
 include_directories("${CMAKE_CURRENT_SOURCE_DIR}")
 include_directories(${CMAKE_CURRENT_BINARY_DIR})
 
-file(GLOB SOCCER_UIS "${CMAKE_CURRENT_SOURCE_DIR}/ui/*.ui")
+file(GLOB SOCCER_UIS "${CMAKE_CURRENT_SOURCE_DIR}/ui/qt/*.ui")
 qt5_wrap_ui(SOCCER_UIS ${SOCCER_UIS})
 
-qt5_add_resources(SOCCER_RSRC ui/main_icons.qrc)
+qt5_add_resources(SOCCER_RSRC ui/qt/main_icons.qrc)
 
 
 # Boost Python
@@ -168,7 +168,7 @@ if(APPLE)
     include(BundleUtilities)
 
     # app icon
-    set(OSX_ICON_FILES ${CMAKE_CURRENT_SOURCE_DIR}/ui/SoccerBall.icns)
+    set(OSX_ICON_FILES ${CMAKE_CURRENT_SOURCE_DIR}/ui/qt/SoccerBall.icns)
     set_source_files_properties(${OSX_ICON_FILES} PROPERTIES MACOSX_PACKAGE_LOCATION Resources)
     set(MACOSX_BUNDLE_ICON_FILE SoccerBall)
 
@@ -191,8 +191,8 @@ qt5_use_modules(soccer Widgets Xml Core OpenGL Network Svg)
 target_link_libraries(soccer robocup)
 
 # build the 'log_viewer' program
-qt5_add_resources(LOG_VIEWER_RSRC ui/log_icons.qrc)
-qt5_wrap_ui(LOG_VIEWER_UI ui/LogViewer.ui)
+qt5_add_resources(LOG_VIEWER_RSRC ui/qt/log_icons.qrc)
+qt5_wrap_ui(LOG_VIEWER_UI ui/qt/LogViewer.ui)
 set(LOG_VIEWER_SRC
     "LogViewer.cpp"
 )
