+++ soccer/ui/BatteryWidget.cpp
