+++ soccer/ui/RobotWidget.cpp
