+++ soccer/ui/RobotWidget.hpp
