+++ soccer/ui/SimFieldView.hpp
@@ -6,7 +6,8 @@
 #include <protobuf/grSim_Commands.pb.h>
 #include <protobuf/grSim_Packet.pb.h>
 #include <protobuf/grSim_Replacement.pb.h>
-#include <FieldView.hpp>
+
+#include "FieldView.hpp"
 #include <QUdpSocket>
 
 class SimFieldView : public FieldView {
