+++ soccer/ui/StripChart.cpp
