+++ soccer/ui/StripChart.hpp
