+++ soccer/ui/qt/ConfigFileTab.ui
