+++ soccer/ui/qt/RefereeTab.ui
