+++ soccer/ui/qt/SoccerBall.icns
