+++ soccer/ui/qt/SoccerBall.png
