+++ soccer/ui/qt/log_icons.qrc
