+++ soccer/ui/qt/main_icons.qrc
