+++ soccer/ui/qt/running.png
