+++ soccer/ui/qt/themes/1337h4x0r.stylesheet
