+++ soccer/ui/qt/themes/QTDark.stylesheet
