+++ soccer/ui/qt/themes/darculized.stylesheet
