+++ soccer/ui/qt/themes/nyan.stylesheet
